Implement TotalMembers on redis repo

+++ leaderboard/repository.go
@@ -9,5 +9,5 @@ type Repository interface {
 	InsertUserScore(ctx context.Context, username string, score float64) error
 	GetUserRank(ctx context.Context, username string) (int64, error)
 	GetUserScore(ctx context.Context, username string) (float64, error)
-	TotalMembers(ctx context.Context) (int, error)
+	TotalMembers(ctx context.Context) (int64, error)
 }
